Rename client to editor
