+++ editor/src/@assets/favicon.ico
