+++ editor/src/renderer.ts
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 
-import App from "@/client/src/views/App.vue";
+import App from "@/editor/src/views/App.vue";
 import { vuetify } from "@/plugins/vuetify";
 
 createApp(App).use(vuetify).mount("#app");
