+++ editor/src/views/App.vue
