+++ editor/tests/App.test.ts
@@ -1,7 +1,7 @@
 import { shallowMount } from "@vue/test-utils";
 import { test } from "vitest";
 
-import App from "@/client/src/views/App.vue";
+import App from "@/editor/src/views/App.vue";
 
 test("Mounted", ({ expect }) => {
   using wrapper = shallowMount(App);
