+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/client/src/renderer.ts"></script>
+    <script type="module" src="/editor/src/renderer.ts"></script>
   </body>
 </html>
