+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["env.d.ts", "common", "client", "plugins", "type_extensions"],
+  "include": ["env.d.ts", "common", "editor", "plugins", "type_extensions"],
   "compilerOptions": {
     "allowImportingTsExtensions": true,
     "baseUrl": ".",
